day4 re-learn & re-practice

+++ day3-depth/modules/invoice.py
@@ -5,5 +5,5 @@ from modules.status import Status
 class Invoice:
     invoice_id : int
     customer_name : str
-    total : float
+    total : int
     status : Status
